Refactor: move AHK template building to its own class.
