+++ Assets/Scripts/System/Classes/AhkBuilder.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 59209f458585f4c20a337cc10742a037
+timeCreated: 1532733152
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
