+++ Assets/Scripts/System/Classes/Game.cs
@@ -43,7 +43,7 @@ public class Game
 
     public bool voidGame = false;
 
-    private JSONNode savedMetadata;
+    public JSONNode savedMetadata;
 
     /// <summary>
     /// The game only takes in a directory handed down by the Playlist class, it then finds all the relevant information for the Game
@@ -182,14 +182,9 @@ public class Game
     private string findFirstImage(string ext) {
         try {
             return Directory.GetFiles(directory.FullName, "*." + ext)[0];
-        } catch (System.IndexOutOfRangeException e) {
+        } catch (System.IndexOutOfRangeException) {
             return null;
         }
-    }
-
-    private string GetExecutablePath() {
-        //Find the .exe in the directory and save a reference
-        return Path.Combine(directory.FullName, executable);
     }
 
     private string GetGameNameFromFolderName() {
@@ -233,52 +228,9 @@ public class Game
     /// and put them in the same folder as the game.
     /// </summary>
     public void BuildHelperScripts() {
-        GM.Instance.logger.Info("GAME: Create scripts for game " + name);
-
-        string newAHKfile = "";
-        string execFile = Path.GetFileName(executable);
-
-        switch (gameType) {
-
-            case Game.GameType.PICO8:
-
-                string newJS = Resources.Load<TextAsset>("Pico8Launcher").text;
-                newJS = newJS.Replace("{{{PATH_TO_HTML}}}", executable.Replace("\\", "\\\\"));
-                WriteStringToFile(newJS, "Pico8Launcher.js");
-
-                newAHKfile = Resources.Load<TextAsset>("AHK_templates/Pico8GameTemplate").text;
-                newAHKfile = newAHKfile.Replace("{GAME_PATH}", GM.Instance.options.dataPath + "/Options/Pico8/nw.exe");
-                break;
-
-            case Game.GameType.FLASH:
-
-                newAHKfile = Resources.Load<TextAsset>("AHK_templates/FlashGameTemplate").text;
-
-                newAHKfile = newAHKfile.Replace("{GAME_PATH}", executable);
-                newAHKfile = newAHKfile.Replace("{GAME_NAME}", name);
-                break;
-
-            default:
-                newAHKfile = Resources.Load<TextAsset>("AHK_templates/ExeGameTemplate").text;
-
-                newAHKfile = newAHKfile.Replace("{GAME_PATH}", executable);
-                newAHKfile = newAHKfile.Replace("{GAME_NAME}", name);
-                break;
-        }
-
-        //Things needed for every Launcher Script
-
-        //Replace variables
-        newAHKfile = newAHKfile.Replace("{GAME_FILE}", execFile);
-        newAHKfile = newAHKfile.Replace("{DEBUG_OUTPUT}", "true"); // TODO make this configurable
-        newAHKfile = newAHKfile.Replace("{IDLE_TIME}", "" + GM.Instance.options.runnerSecondsIdle);
-        newAHKfile = newAHKfile.Replace("{IDLE_INITIAL}", "" + GM.Instance.options.runnerSecondsIdleInitial);
-        newAHKfile = newAHKfile.Replace("{ESC_HOLD}", "" + GM.Instance.options.runnerSecondsESCHeld);
-
-        newAHKfile = insertKeyMapping(newAHKfile);
-
-        //Delete old file and write to new one
-        WriteStringToFile(newAHKfile, "RunGame.ahk");
+        AhkBuilder ahk = new AhkBuilder(this);
+        ahk.compile();
+        ahk.write();
     }
 
     /// <summary>
@@ -295,115 +247,12 @@ public class Game
         }
     }
 
-
-    private string insertKeyMapping(string ahkFile) {
-        string keymap = "";
-        JSONNode parsedBindings = getKeyBindings();
-        ArrayList gameKeys = allGameKeys(parsedBindings);
-
-        // Write keys for players we have.
-        for(int pNum = 1; pNum <= savedMetadata["max_players"].AsInt; pNum++) {
-            JSONNode playerKeys;
-
-            try {
-                playerKeys = parsedBindings[pNum.ToString()];
-            } catch (System.NullReferenceException) {
-                break;
-            }
-
-            foreach(string control in KeyBindings.CONTROLS) {
-                KeyCode key = GM.Instance.options.keys.GetKey(pNum, control);
-                string launcherKey = GM.Instance.options.keyTranslator.toAHK(key);
-                string gameKey = playerKeys[control];
-
-                if (gameKey == null) { // this shouldn't happen, but just in case.
-                    gameKey = "return";
-                }
-
-                keymap += (launcherKey + "::" + gameKey + "\n");
-            }
-        }
-
-        // Write keys for players we don't have (e.g., players 3 & 4 on 2-player game)
-        for(int pNum = savedMetadata["max_players"].AsInt + 1; pNum <= 4; pNum++) {
-            foreach(string control in KeyBindings.CONTROLS) {
-                KeyCode key = GM.Instance.options.keys.GetKey(pNum, control);
-                string launcherKey = GM.Instance.options.keyTranslator.toAHK(key);
-
-                if (!gameKeys.Contains(launcherKey))
-                    keymap += (launcherKey + "::return\n");
-            }
-        }
-
-        return ahkFile.Replace("{KEYMAP}", keymap);
-    }
-
-    private ArrayList allGameKeys(JSONNode bindings) {
-        ArrayList keys = new ArrayList();
-
-        for(int pNum = 1; pNum <= 4; pNum++) {
-            foreach(string control in KeyBindings.CONTROLS) {
-                string key = bindings[pNum.ToString()][control];
-
-                if (key != null)
-                    keys.Add(key);
-            }
-        }
-
-        return keys;
-    }
-
-    private JSONNode getKeyBindings() {
-        string tmpl = savedMetadata["keys"]["template"];
-        JSONNode bindings = null;
-
-        string tmplFile = Path.Combine(GM.Instance.options.defaultOptionsPath, "keymap_templates.json");
-        JSONNode bindingTemplates = GM.Instance.data.LoadJson(tmplFile);
-
-        if (tmpl == null) {
-            if (savedMetadata["keys"]["bindings"] == null) {
-                GM.Instance.logger.Debug("No key binding info provided for " + name + ". Using defaults.");
-                tmpl = "default";
-            } else {
-                tmpl = "custom";
-            }
-        }
-
-        switch(tmpl) {
-            case "custom":
-                GM.Instance.logger.Debug("Loading custom key bindings for " + name);
-                bindings = savedMetadata["keys"]["bindings"];
-                break;
-
-            case "default":
-            case "legacy":
-            case "flash":
-            case "pico8":
-                GM.Instance.logger.Debug("Loading " + tmpl + " bindings from tmplFile: " + tmplFile);
-                bindings = bindingTemplates[tmpl];
-                break;
-
-            default:
-                GM.Instance.logger.Error("Invalid key template type '" + tmpl + "' for " + name + ". (Using default.) Valid templates are 'default', 'legacy', 'pico8', 'custom'.");
-                GM.Instance.logger.Debug("Loading " + tmpl + " bindings from tmplFile: " + tmplFile);
-                bindings = bindingTemplates["default"];
-                break;
-        }
-
-        // Remove controls for players that don't exist.
-        for (int p = savedMetadata["max_players"].AsInt + 1; p <= 4; p++) {
-            bindings.Remove(p.ToString());
-        }
-
-        return bindings;
-    }
-
     /// <summary>
     /// Writes the text to the filename in the Game directory.
     /// </summary>
     /// <param name="text">The text to encode into the file.</param>
     /// <param name="fileName">The name of the file.</param>
-    private void WriteStringToFile(string text, string fileName) {
+    public void WriteStringToFile(string text, string fileName) {
         string behaviour = GM.Instance.options.O["launcher"]["writeScripts"];
         if (behaviour == null)
             behaviour = "always"; // backwards-compatible default behavior
